mensaje de aviso en caso de que no hayan ventas hecho
